Implemented blood drip animation as a game mechanic.
The game will now detect if blood has hit a wall and spawn the blood drip animation to play. There is one problem however, the quad for the animation is rotated differently depending on which side of the cube the quad lands on, meaning on one side the quad will be rotated correctly and on the opposite side the quad might be upside down. This will be fixed tomorrow.